refactor: 인터페이스 구현체 파일 통합
