--- src/main/kotlin/racingcar/domain/RandomNumberGenerator.kt
@@ -1,9 +1,0 @@
-package racingcar.domain
-
-import kotlin.random.Random
-
-object RandomNumberGenerator : NumberGenerator {
-    override fun generate(from: Int, until: Int): Int {
-        return Random.nextInt(from, until)
-    }
-}
